Moving the new matrix into the right file (#6)
